% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.data.frame.matrix_handle}
\alias{as.data.frame.matrix_handle}
\title{S3 generic for converting a \code{matrix_handle} into a \code{data.frame}}
\usage{
\method{as.data.frame}{matrix_handle}(x)
}
\arguments{
\item{x}{\code{matrix_handle}}
}
\description{
S3 generic for converting a \code{matrix_handle} into a \code{data.frame}
}
