% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.matrix.CaliperMatrix}
\alias{as.matrix.CaliperMatrix}
\title{S3 method for bringing caliper matrices into R matrix format}
\usage{
\method{as.matrix}{CaliperMatrix}(x, ..., core = NULL)
}
\arguments{
\item{x}{A \code{CaliperMatrix} object}

\item{...}{Additional arguments passed to \code{as.matrix}.}

\item{core}{\code{string} Optional name of core to convert. Defaults to
first core.}
}
\value{
an R matrix
}
\description{
Converts a Caliper matrix into R format allowing it to be worked on in the R
environment.
}
\details{
By default, the first core of the Caliper matrix is converted into an R
matrix. To select a different one, use the \code{core} argument like so:
\code{as.matrix(x, core = "second core")}.
}
