% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.matrix.matrix_handle}
\alias{as.matrix.matrix_handle}
\title{S3 generic for bringing caliper matrices into R matrix format}
\usage{
\method{as.matrix}{matrix_handle}(x, core = NULL)
}
\arguments{
\item{x}{A \code{matrix_currency} object}

\item{core}{\code{string} Name of core to convert. If not provided, the first
core will be used.}
}
\description{
S3 generic for bringing caliper matrices into R matrix format
}
