% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.matrix.MatrixCurrency}
\alias{as.matrix.MatrixCurrency}
\title{Generic to convert MatrixCurrency class to an R matrix}
\usage{
\method{as.matrix}{MatrixCurrency}(x, ...)
}
\arguments{
\item{x}{\code{MatrixCurrency} One of the cores of the \code{CaliperMatrix}
class.}

\item{...}{Other arguments passed to \code{as.matrix}}
}
\value{
An R matrix.
}
\description{
Converts a Caliper matrix into R format allowing it to be worked on in the R
environment.
}
