% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper.R
\docType{package}
\name{caliper}
\alias{caliper}
\alias{caliper-package}
\title{caliper: Communicate with Caliper software from R}
\description{
The caliper package provides functions to make it easier to communicate
with Caliper software over COM.
}
