% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CaliperClass}
\alias{CaliperClass}
\title{R class representing GISDK objects}
\description{
This class simplifies working with GISDK objects over COM from R.
}
\section{Overview}{


When created, an object of this class contains a COM pointer to a sister
object in Caliper software. In addition to this reference, contained in the
\code{ref} field, there are several other fields and methods to facilitate
object manipulation. Finally, the \code{$} operator has been overloaded for
this class to make code easier to write.
}

\section{Object fields}{


A \code{CaliperClass} object has the following fields/attributes of interest:
\describe{
  \item{g_class_name}{
    The class name of the underlying GISDK object (e.g. "NLM.Model").
  }
  \item{ref}{
    This is a COM pointer and represents the object created in a Caliper
    program.
  }
  \item{info}{
    This is a simple R list that can be used to see the fields and methods
    of the underlying GISDK object.
  }
}
}

\section{Object methods}{


A \code{CaliperClass} object has several methods. While they are listed
below for completeness, you likely won't need to use the methods directly
(see examples).

\describe{
  \item{apply_gisdk_method}{
    Dispatches one of the GISDK object's methods
  }
  \item{get_gisdk_attribute(attribute)}{
    Gets the value of a named attribute of the GISDK object
  }
  \item{set_gisdk_attribute(attribute)}{
    Sets the value of a named attribute of the GISDK object
  }
}
}

\examples{
\dontrun{
# object creation (easier to use caliper::CreateObject())
obj <- CaliperClass$new("NLM.Model")

# get info about the GISDK object fields and methods
obj$info

# set an attribute/field label of the GISDK object
# (Does not modify the R object in any way)
obj$Label <- "a logit model"

# run one of the GISDK object's methods: Clear()
# (Does not modify the R object in any way)
obj$Clear()

# object methods can be chained
obj$
 Clear()$
 Read()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CaliperClass$new()}}
\item \href{#method-apply_gisdk_method}{\code{CaliperClass$apply_gisdk_method()}}
\item \href{#method-get_gisdk_attribute}{\code{CaliperClass$get_gisdk_attribute()}}
\item \href{#method-set_gisdk_attribute}{\code{CaliperClass$set_gisdk_attribute()}}
\item \href{#method-clone}{\code{CaliperClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperClass$new(class_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_gisdk_method"></a>}}
\subsection{Method \code{apply_gisdk_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperClass$apply_gisdk_method(method, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_gisdk_attribute"></a>}}
\subsection{Method \code{get_gisdk_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperClass$get_gisdk_attribute(attribute)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_gisdk_attribute"></a>}}
\subsection{Method \code{set_gisdk_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperClass$set_gisdk_attribute(attribute, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
