% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\docType{data}
\name{CaliperMatrix}
\alias{CaliperMatrix}
\title{CaliperMatrix class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CaliperMatrix
}
\description{
This class makes it easier to interact with Caliper matrices in R.
}
\section{Object fields}{


A \code{CaliperMatrix} object has the following fields/attributes of interest:
\describe{
  \item{handle}{
    This is a COM pointer and represents the object in Caliper Software.
  }
  \item{cores}{
    This is a named list of cores. Each element contains a COM pointer to
    a matrix currency in Caliper software.
  }
  \item{indices}{
    This lists the row and column indices available for the matrix.
  }
  \item{row_index}{
    This can be used to retrieve or set the current row index.
  }
  \item{column_index}{
    This can be used to retrieve or set the current column index.
  }
}

See the vignette "Using caliper" for more details on interacting with
matrices. This includes info on S3 methods for bringing them into R formats
like \code{data.frames} and \code{matrices}.
}

\keyword{datasets}
