% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperr.R
\docType{package}
\name{caliperr}
\alias{caliperr}
\alias{caliperr-package}
\title{caliperr: Communicate with Caliper software from R}
\description{
The caliperr package provides functions to make it easier to communicate
with Caliper software over COM.
}
