% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$.CaliperClass}
\alias{$.CaliperClass}
\title{S3 method for calling \code{CaliperClass} object methods}
\usage{
\method{$}{CaliperClass}(x, name)
}
\arguments{
\item{x}{A \code{CaliperClass} object}

\item{name}{the method to dispatch}
}
\description{
Makes \code{CaliperClass} objects smarter about whether you are calling a
method from the R object or the underlying GISDK object over COM.
}
\details{
If \code{name} is an attribute of the R object (like \code{$info}), then
the value of that attribute is returned. Otherwise, it looks into the fields
and methods of the underlying GISDK object to determine what to do.
}
