% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{$.CaliperMatrix}
\alias{$.CaliperMatrix}
\title{S3 method for calling \code{CaliperMatrix} object attributes}
\usage{
\method{$}{CaliperMatrix}(x, name)
}
\arguments{
\item{x}{A \code{CaliperMatrix} object}

\item{name}{the method to dispatch}
}
\description{
Makes it easier to call a matrix core directly using \code{matrix$core_name}.
}
\details{
If \code{name} is an attribute of the R object (like \code{$info}), then
the value of that attribute is returned. Otherwise, it looks into the fields
and methods of the underlying GISDK object to determine what to do.
}
