% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaliperR-class.R
\name{$.CaliperR}
\alias{$.CaliperR}
\title{S3 method for calling \code{CaliperR} object methods}
\usage{
\method{$}{CaliperR}(x, name, ..., process_result = TRUE)
}
\arguments{
\item{x}{A \code{CaliperR} object}

\item{name}{the method to dispatch}
}
\description{
Makes \code{CaliperR} objects smarter about whether you are calling a
method from the R object or the underlying GISDK library.
}
\details{
If \code{name} is a method of the R object, then it is executed. Otherwise,
it tries to execute a GISDK function of the same name.
}
