% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$.GisdkObject}
\alias{$.GisdkObject}
\title{S3 generic for calling GisdkObject methods}
\usage{
\method{$}{GisdkObject}(x, name)
}
\arguments{
\item{x}{A \code{GisdkObject}}

\item{name}{the method to dispatch}
}
\description{
Makes GisdkObjects smarter about whether you are calling a method from the R
object or the underlying GISDK object over COM.
}
