% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$<-.CaliperClass}
\alias{$<-.CaliperClass}
\title{S3 method for assigning \code{CaliperClass} object attributes}
\usage{
\method{$}{CaliperClass}(x, name) <- value
}
\arguments{
\item{x}{A \code{CaliperClass} object}

\item{name}{the attribute to assign}

\item{value}{the value to be assigned}
}
\description{
Makes \code{CaliperClass} objects smarter about whether you are assigning a
value to the R object or the underlying GISDK object over COM.
}
\details{
If \code{name} is an attribute of the R object (like \code{$info}), then
the value is assigned to that attribute. Otherwise, it looks into the GISDK
object attributes and will set that value.
}
