% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{$<-.CaliperMatrix}
\alias{$<-.CaliperMatrix}
\title{Send matrix data back to Caliper software}
\usage{
\method{$}{CaliperMatrix}(x, name) <- value
}
\arguments{
\item{x}{A \code{CaliperMatrix} object}

\item{name}{the attribute to assign}

\item{value}{the value to be assigned}
}
\description{
Standard assignment would simply overwrite the matrix currency with whatever
data you assigned. This allows for properly formed matrix data to be sent
back to Caliper software and update the matrix currency.
}
