% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$<-.GisdkObject}
\alias{$<-.GisdkObject}
\title{S3 generic for assigning GisdkObject attributes}
\usage{
\method{$}{GisdkObject}(x, name) <- value
}
\arguments{
\item{x}{A \code{GisdkObject}}

\item{name}{the attribute to assign}

\item{value}{the value to be assigned}
}
\description{
Makes GisdkObjects smarter about whether you are assigning a value to the R
object or the underlying GISDK object over COM.
}
