% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{checkIfValidBinFile}
\alias{checkIfValidBinFile}
\title{Checks if filename is a valid .bin file}
\usage{
checkIfValidBinFile(binFilename)
}
\arguments{
\item{binFilename}{A data table with columns to convert}
}
\description{
Stops the program if the filename is not a valid .bin extension
}
\keyword{internal}
