% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{compile_gisdk}
\alias{compile_gisdk}
\title{Compiles a GISDK script file into a UI}
\usage{
compile_gisdk(rsc_file, ui_file = NULL)
}
\arguments{
\item{rsc_file}{\code{string} Script file path.}

\item{ui_file}{\code{string} Optional output location of the compiled code.
By default, it will use the same file name as the \code{rsc_file}, but with
a .dbd extension.}
}
\value{
The file path of the compiled '.dbd' file. Like other Caliper databases,
  the compiled UI is made up of several related files.
}
\description{
Compiles a GISDK script file into a UI
}
