% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperr.R
\name{convert_to_gisdk_null}
\alias{convert_to_gisdk_null}
\title{Converts R's \code{NA} and \code{NULL} to \code{NA_complex_)}}
\usage{
convert_to_gisdk_null(arg)
}
\value{
Returns the argument passed in with any \code{NA/NULL} converted.
}
\description{
\code{NA_complex_} is understood by GISDK/C++ as null.
}
\keyword{internal}
