% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{convert_to_named_list}
\alias{convert_to_named_list}
\title{Used to convert GISDK named arrays to R's named lists.}
\usage{
convert_to_named_list(nested_list)
}
\arguments{
\item{nested_list}{The list object that results whenever GISDKs named arrays
are passed through COM.}
}
\value{
A named list
}
\description{
When a GISDK named array comes across COM, it has a specific format that
isn't easy to use. This converts it into an R named list.
}
\keyword{internal}
