% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{convertColClasses}
\alias{convertColClasses}
\title{Coerces data.frame columns to the specified classes}
\usage{
convertColClasses(dataTable, colClasses)
}
\arguments{
\item{dataTable}{A data table with columns to convert}

\item{colClasses}{A vector specifying the classes of each column}
}
\description{
Coerces data.frame columns to the specified classes
}
\keyword{internal}
