% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{correct_empty_columns}
\alias{correct_empty_columns}
\title{Correct the type for empty columns}
\usage{
correct_empty_columns(df, r_types)
}
\arguments{
\item{df}{\code{data.frame} The data.frame to correct.}

\item{r_types}{\code{list/vector} of data type strings.}
}
\description{
Empty columns read in through various means end up with type = logical.
This function uses the dcbKey to correct them.
}
\keyword{internal}
