% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{create_matrix}
\alias{create_matrix}
\title{Creates an S3 object of class \code{matrix_handle}}
\usage{
create_matrix(file)
}
\arguments{
\item{file}{File name to read (.mtx)}
}
\description{
A \code{matrix_handle} is different from an R matrix. This object represents
all the cores of a Caliper matrix.
}
