% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{create_matrix_cores}
\alias{create_matrix_cores}
\title{Internal function used to create currency pointers for all cores in a matrix}
\usage{
create_matrix_cores(mh, ri = NULL, ci = NULL)
}
\arguments{
\item{mh}{\code{COMIDispatch} A pointer to the GISDK matrix handle}

\item{ri}{\code{string} Name of the row index. (NULL for base index)}

\item{ci}{\code{string} Name of the column index. (NULL for base index)}
}
\description{
This is the function that creates the list of COM pointers (one for each core)
that is stored in a \code{CaliperMatrix} object.
}
\keyword{internal}
