% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperr.R
\name{create_named_array}
\alias{create_named_array}
\title{Used internally to convert R's named vectors/lists to GISDK named arrays.}
\usage{
create_named_array(named_list)
}
\arguments{
\item{named_list}{A named list (or vector)}
}
\value{
a pointer object that GISDK will interpret as a named array
}
\description{
Used internally to convert R's named vectors/lists to GISDK named arrays.
}
\keyword{internal}
