% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{create_unique_view_name}
\alias{create_unique_view_name}
\title{Creates a view name guaranteed to be unique in the current Caliper session}
\usage{
create_unique_view_name()
}
\description{
Creates a view name guaranteed to be unique in the current Caliper session
}
\keyword{internal}
