% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{create_view}
\alias{create_view}
\title{Creates a new Caliper view with data from a data.frame}
\usage{
create_view(df)
}
\description{
Creates a new Caliper view with data from a data.frame
}
\keyword{internal}
