% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateObject}
\alias{CreateObject}
\title{Creates an R object that reprents a GISDK object}
\usage{
CreateObject(class_name, ...)
}
\arguments{
\item{class_name}{The class name of the GISDK object to create.}

\item{...}{Any additional arguments to use during GISDK object creation.}
}
\description{
A simple function that merely abstracts Rs creation language for R6 objects
to make it look more like GISDK. See \code{\link{CaliperClass}} for details
on the class and how to work with objects.
}
\examples{
\dontrun{
obj <- CreateObject("NLM.Model")
}
}
