% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{df_from_view}
\alias{df_from_view}
\title{Converts a Caliper view into an R data.frame}
\usage{
df_from_view(view_name, set_name = NULL)
}
\arguments{
\item{view_name}{The name of the view open in Caliper software.}

\item{set_name}{An optional set name can be provided to only return
records in a selection set from the view.}
}
\description{
Converts a Caliper view into an R data.frame
}
