% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{df_to_view}
\alias{df_to_view}
\title{Converts a data.frame into a view. Updates an existing view or creates
a new one as needed.}
\usage{
df_to_view(df, view_name = NULL, set_name = NULL)
}
\description{
Converts a data.frame into a view. Updates an existing view or creates
a new one as needed.
}
