% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{disconnect}
\alias{disconnect}
\title{Close the COM connection to Caliper software and kills the process}
\usage{
disconnect()
}
\value{
Nothing.
}
\description{
Close the COM connection to Caliper software and kills the process
}
