% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{get_package_variable}
\alias{get_package_variable}
\title{Gets the value of a \code{caliperR} package-wide variable}
\usage{
get_package_variable(package_variable)
}
\arguments{
\item{package_variable}{The package variable to set}
}
\description{
Gets the value of a \code{caliperR} package-wide variable
}
\keyword{internal}
