% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{getByteLength}
\alias{getByteLength}
\title{Return the byte length for each R data type}
\usage{
getByteLength(binDataCol)
}
\arguments{
\item{binDataCol}{The data table column for which data type to return}
}
\value{
byte length for each R data type
}
\description{
Return the byte length for each R data type
}
\keyword{internal}
