% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{getDisplayLength}
\alias{getDisplayLength}
\title{Return the display length for each R data type}
\usage{
getDisplayLength(binDataCol)
}
\arguments{
\item{binDataCol}{The data table column for which data type to return}
}
\value{
A default character length for each field type
}
\description{
Return the display length for each R data type
}
\keyword{internal}
