% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GisdkClass.R
\name{GisdkClass}
\alias{GisdkClass}
\title{R class that gives access to GISDK functions}
\description{
R class that gives access to GISDK functions

R class that gives access to GISDK functions
}
\details{
This class makes it easy to call any arbitrary GISDK function from R.

Performs multiple dispatch based on the name of the method called. If the
named is one of a few caliperR functions, those are executed. If not, then
then the name is assumed to represent a GISDK function and is called over
COM.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GisdkClass$new()}}
\item \href{#method-RunMacro}{\code{GisdkClass$RunMacro()}}
\item \href{#method-CreateObject}{\code{GisdkClass$CreateObject()}}
\item \href{#method-SetAlternateInterface}{\code{GisdkClass$SetAlternateInterface()}}
\item \href{#method-GetInterface}{\code{GisdkClass$GetInterface()}}
\item \href{#method-clone}{\code{GisdkClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
On creation, attempts to connect to Caliper software if not
already connected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{GisdkClass} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunMacro"></a>}}
\if{latex}{\out{\hypertarget{method-RunMacro}{}}}
\subsection{Method \code{RunMacro()}}{
Class implementation of \code{caliperR::RunMacro()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$RunMacro(macro_name, ..., process_result = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{macro_name}}{\code{string} The GISDK macro name to run.}

\item{\code{...}}{The arguments to pass to the GISDK macro.}

\item{\code{process_result}}{\code{boolean} Whether to attempt to convert GISDK
data types to R. Defaults to true.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateObject"></a>}}
\if{latex}{\out{\hypertarget{method-CreateObject}{}}}
\subsection{Method \code{CreateObject()}}{
Class implementation of \code{caliperR::CreateObject()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$CreateObject(class_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class_name}}{\code{string} The name of the GISDK object to create.}

\item{\code{...}}{Any arguments to be passed to the object during creation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetAlternateInterface"></a>}}
\if{latex}{\out{\hypertarget{method-SetAlternateInterface}{}}}
\subsection{Method \code{SetAlternateInterface()}}{
Class implementation of \code{caliperR::SetAlternateInterface()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$SetAlternateInterface(ui_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ui_file}}{\code{string} File path to the compiled UI file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetInterface"></a>}}
\if{latex}{\out{\hypertarget{method-GetInterface}{}}}
\subsection{Method \code{GetInterface()}}{
Class implementation of \code{caliperR::GetInterface()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$GetInterface()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GisdkClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
