% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{GisdkObject}
\alias{GisdkObject}
\title{R class representing GISDK objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GisdkObject
}
\description{
This class simplifies working with GISDK objects over COM from R.
}
\examples{
\dontrun{
# object creation
obj <- GisdkObject$new("NLM.Model")

# get info about the GISDK object
obj$g_info

# run one of the object's methods
obj$Clear()

# object methods can be chained
obj$
 Clear()$
 Read()
}
}
\keyword{datasets}
