% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{is_gisdk_named_array}
\alias{is_gisdk_named_array}
\title{Checks if an object is a GISDK named array}
\usage{
is_gisdk_named_array(object)
}
\arguments{
\item{object}{The object to be tested}
}
\value{
\code{boolean} TRUE/FALSE
}
\description{
When returned over COM, these arrays have a very specific format. This
function is used in conjunction with \code{\link{convert_to_named_list}} in
\code{\link{process_gisdk_result}} to identify and convert GISDK named arrays
into R's named lists.
}
\keyword{internal}
