% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{make_MatrixCurrency}
\alias{make_MatrixCurrency}
\title{Creates a matrix_currency object}
\usage{
make_MatrixCurrency(x)
}
\arguments{
\item{x}{\code{COMIDispatch} Pointer to a GISDK matrix currency.}
}
\value{
A \code{MatrixCurrency} object.
}
\description{
This is used to wrap a COMIDispatch pointer (representing a GISDK matrix
currency).
}
\keyword{internal}
