% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\docType{class}
\name{MatrixCurrency-class}
\alias{MatrixCurrency-class}
\title{The matrix currency class}
\description{
This is a simple S4 class that wraps a COMIDispatch pointer. That pointer
points to a matrix currency in Caliper software.
}
\keyword{internal}
