% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{print.matrix_currency_temp}
\alias{print.matrix_currency_temp}
\title{S3 generic for printing a matrix currency}
\usage{
\method{print}{matrix_currency_temp}(x, ...)
}
\arguments{
\item{x}{A \code{matrix_currency} object}

\item{...}{Additional options}
}
\description{
S3 generic for printing a matrix currency
}
