% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{print.matrix_handle_temp}
\alias{print.matrix_handle_temp}
\title{S3 generic for printing a \code{matrix_handle}}
\usage{
\method{print}{matrix_handle_temp}(x, ...)
}
\arguments{
\item{x}{A \code{matrix_handle} object}

\item{...}{Additional options}
}
\description{
S3 generic for printing a \code{matrix_handle}
}
