% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{process_gisdk_result}
\alias{process_gisdk_result}
\title{Converts GISDK output into R structures}
\usage{
process_gisdk_result(result)
}
\arguments{
\item{result}{A returned value from Caliper software.}
}
\description{
Simple objects pass through the COM interface as usable R structures.
GISDK arrays, for example, become R vectors. Other things, like GISDK
vectors, come across as pointers to the object inside the Caliper
process. This function attemps to coerce those pointers into usable
R data structures.
}
\keyword{internal}
