% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{read_bin}
\alias{read_bin}
\title{Read a bin file}
\usage{
read_bin(binFilename, returnDnames = FALSE)
}
\arguments{
\item{binFilename}{\code{string} The bin filename to read}

\item{returnDnames}{\code{bool} To return display names (if present in bin)}
}
\value{
The data table read from the bin located at binFilename
}
\description{
Read the data in binfilename and return a data table. Any field descriptions
will become column labels (use \code{View()} or \code{Hmisc::label()} to see
them).
}
