% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{read_bin_without_com}
\alias{read_bin_without_com}
\title{Read the bin matrix into a data table without a COM connection}
\usage{
read_bin_without_com(binData, binMatrix, dcbKey, TcDataType, nRows, nCols)
}
\arguments{
\item{binData}{Empty matrix with correct dimensions.}

\item{binMatrix}{The raw bytes converted to a matrix.}

\item{dcbKey}{Data frame of column info.}

\item{TcDataType}{The TransCAD data type.}

\item{nRows}{Number of rows in the FFB.}

\item{nCols}{Number of columns in the FFB.}
}
\value{
The modified binData data table
}
\description{
Helper function to \code{\link{read_bin}}. If Caliper software is not
installed, this function reads and translates the bytes directly.
}
\keyword{internal}
