% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{readFfb}
\alias{readFfb}
\title{Read the bin matrix into a data table}
\usage{
readFfb(binData, binMatrix, dcbKey, TcDataType, nRows, nCols)
}
\value{
The modified binData data table
}
\description{
Where all the heavy lifting is done
Reads data table column by column after the data is read
into a binary matrix
}
\keyword{internal}
