% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{RNaToTcMiss}
\alias{RNaToTcMiss}
\title{Convert R NAs to TransCAD missing values}
\usage{
RNaToTcMiss(value, typeChar)
}
\arguments{
\item{value}{Value of the data table entry to convert}

\item{typeChar}{String denoting the TC type}
}
\value{
Either the value passed in or a TC missing integer/real number
}
\description{
Convert R NAs to TransCAD missing values
}
\keyword{internal}
