% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{RTypeToTcType}
\alias{RTypeToTcType}
\title{Convert R datatype strings to TransCAD type string}
\usage{
RTypeToTcType(typeChar)
}
\arguments{
\item{typeChar}{String denoting the R type}
}
\value{
The correponding TC type
}
\description{
Convert R datatype strings to TransCAD type string
}
\keyword{internal}
