% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperr.R
\name{run_macro}
\alias{run_macro}
\title{Runs a macro (function) in GISDK}
\usage{
run_macro(macro_name = NULL, ...)
}
\arguments{
\item{macro_name}{\code{string} Name of the GISDK macro to run}

\item{...}{Used to pass arguments to the GISDK macro}

\item{ui}{\code{string} Optional. Can be used to point to a custom UI
compiled by the user. Defaults to \code{Sys.getenv("CALIPER_UI")}.}
}
\description{
Runs a macro (function) in GISDK
}
\examples{
\dontrun{
# These won't work unless Caliper software is installed.
run_macro("G30 Tutorial Folder")
run_macro("add", ui = ui_path, 1, 2)
# 3
run_macro("parse opts array", ui = ui_path, list("one" = 1, "two" = 2))
# "The first option name is one. The first option value is 1."
}
}
