% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{RunFunction}
\alias{RunFunction}
\title{Runs a GISDK function}
\usage{
RunFunction(macro_name, ..., process_result = TRUE)
}
\arguments{
\item{macro_name}{\code{string} Name of the GISDK function to run}

\item{...}{Used to pass arguments to the GISDK function}

\item{process_result}{\code{boolean} Whether to attempt to process the
result into a native R format.}
}
\description{
GISDK functions are core functions (like \code{OpenTable()}) that are called
in Caliper software without using \code{RunMacro()}.
}
\details{
To run GISDK macros (like \code{RunMacro("G30 Tutorial Folder")}) see
\code{\link{RunMacro}}.
}
\examples{
\dontrun{
# These won't work unless Caliper software is installed.
table_name <- RunFunction("OpenTable", "airports", "ffb", list(paste0(folder, "airports.bin"), NA))
num_rows <- RunFunction("GetRecordCount", table_name, NA)
num_rows
#> 280
}
}
\keyword{internal}
