% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{RunMacro}
\alias{RunMacro}
\title{Runs a GISDK macro}
\usage{
RunMacro(macro_name, ..., process_result = TRUE)
}
\arguments{
\item{macro_name}{\code{string} Name of the GISDK macro to run}

\item{...}{Used to pass arguments to the GISDK macro}

\item{process_result}{\code{boolean} Whether to attempt to process the
result into a native R format.}
}
\description{
A GISDK macro is a function defined in a GISDK scrip (.rsc). When working in
Caliper software, these are always called in GISDK using the
\code{RunMacro()} function.
}
\details{
To run GISDK functions (like \code{OpenTable()}) see \code{\link{RunFunction}}.
}
\examples{
\dontrun{
# These won't work unless Caliper software is installed.
RunMacro("G30 Tutorial Folder")
RunMacro("add", 1, 2)
#> 3
RunMacro("parse opts array", list("one" = 1, "two" = 2))
#> "The first option name is one. The first option value is 1."
}
}
