% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{set_package_variable}
\alias{set_package_variable}
\title{Sets the value of a package-wide variable}
\usage{
set_package_variable(package_variable, value)
}
\arguments{
\item{package_variable}{The package variable to set}

\item{value}{the value to set with}
}
\description{
\code{caliperR} uses several package-wide variables to enable communication
between functions and simplify function arguments.
}
\keyword{internal}
