% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{SetAlternateInterface}
\alias{SetAlternateInterface}
\title{Change the Caliper UI}
\usage{
SetAlternateInterface(ui_file = NULL)
}
\arguments{
\item{ui_file}{\code{string} File path to the custom UI. If null, will set
the interface back to the default.}
}
\description{
Often, a user will have created their own GISDK functions and compiled them
to a UI file (.dbd). To run them (using \code{\link{RunMacro}}), first use
this function to point to the custom UI. This can also be used to set the
UI back to the default.
}
\details{
To see the current UI, use \code{\link{GetInterface}}.
}
