% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{setupTcCommandPrompt}
\alias{setupTcCommandPrompt}
\title{Configures the environment for TransCAD}
\usage{
setupTcCommandPrompt()
}
\description{
Sets the error to dump.frames to prevent execution halt
Turns off command echo
}
\keyword{internal}
