% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{split_matrix}
\alias{split_matrix}
\title{Splits a matrix by row and column indices}
\usage{
split_matrix(M, list_of_rows, list_of_cols)
}
\arguments{
\item{M}{\code{matrix} A matrix full of raw bytes.}

\item{list_of_rows}{\code{list} The indices of rows to split
(e.g. \code{list(1:4, 5:8)}).}

\item{list_of_cols}{\code{list} The indices of columns to split.
(e.g. \code{list(0:3)}).}
}
\description{
Helper matrix to \code{\link{read_bin_without_com}}.
}
\keyword{internal}
