% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{summary.CaliperMatrix}
\alias{summary.CaliperMatrix}
\title{S3 method for summarizing a \code{CaliperMatrix}}
\usage{
\method{summary}{CaliperMatrix}(object, ...)
}
\arguments{
\item{object}{A \code{CaliperMatrix} object.}

\item{...}{Additional arguments passed to the \code{summary} generic
function.}
}
\description{
S3 method for summarizing a \code{CaliperMatrix}
}
