% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{summary.matrix_handle}
\alias{summary.matrix_handle}
\title{S3 method for summarizing a \code{matrix_handle}}
\usage{
\method{summary}{matrix_handle}(x, ...)
}
\arguments{
\item{x}{\code{matrix_handle}}

\item{...}{Additional arguments (not used)}
}
\description{
S3 method for summarizing a \code{matrix_handle}
}
