% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{TcMissToRNa}
\alias{TcMissToRNa}
\title{Convert TransCAD missing values to R NAs}
\usage{
TcMissToRNa(value, typeChar)
}
\arguments{
\item{value}{Value of the data table entry to convert}

\item{typeChar}{String denoting the TC type}
}
\value{
Either the value passed in or an NA
}
\description{
Convert TransCAD missing values to R NAs
}
\keyword{internal}
