% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{TcTypeToRType}
\alias{TcTypeToRType}
\title{Convert TransCAD type string to R datatype strings}
\usage{
TcTypeToRType(typeChar)
}
\arguments{
\item{typeChar}{String denoting the TC type}
}
\value{
The correponding R type
}
\description{
Convert TransCAD type string to R datatype strings
}
\keyword{internal}
