% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{TcTypeToRType2}
\alias{TcTypeToRType2}
\title{Convert TransCAD type string to R datatype strings}
\usage{
TcTypeToRType2(typeChar)
}
\arguments{
\item{typeChar}{String denoting the TC type}
}
\value{
The correponding R type
}
\description{
Unlike \code{\link{TcTypeToRType}}, this function works on the Caliper types
returned by GISDKs \code{GetFieldType}, which are words like "Integer" instead
of characters like "I".
}
\keyword{internal}
