% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{write_bin}
\alias{write_bin}
\title{Write data table into bin file}
\usage{
write_bin(binData, binFilename, description = "", dnames = NA, n2i = TRUE)
}
\arguments{
\item{binData}{\code{data.frame} The data table to write}

\item{binFilename}{\code{string} The file in which to write}

\item{description}{\code{string} A description of the binData}

\item{dnames}{\code{Named vector} of display names where each name is a
column name and each value is the display name to use. For example:
\code{c("hh" = "Households")} will assign the display name "Households" to
column "hh".}

\item{n2i}{\code{bool} Whether to try to convert numerics to integers.
This leads to integer columns in the bin (instead of real), which
are smaller and easier to read. Defaults to \code{TRUE}. Conversion
does not happen if the column contains decimal values.}
}
\description{
Write the data table into a TransCAD binary file (.bin)
given the filename and description. If the table has
column labels, those will become field descriptions.
}
