% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.data.frame.CaliperMatrix}
\alias{as.data.frame.CaliperMatrix}
\title{S3 method for converting a \code{CaliperMatrix} into a \code{data.frame}}
\usage{
\method{as.data.frame}{CaliperMatrix}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{\code{CaliperMatrix}}

\item{row.names}{See \code{as.data.frame}}

\item{optional}{See \code{as.data.frame}}

\item{...}{additional arguments passed to \code{as.data.frame}}
}
\value{
Returns a \code{data.frame} with \code{from} and \code{to} columns
  along with a column for each core of the matrix. The \code{data.frame}
  respects the current row and column indices.
}
\description{
S3 method for converting a \code{CaliperMatrix} into a \code{data.frame}
}
