% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.matrix.CaliperMatrix}
\alias{as.matrix.CaliperMatrix}
\title{S3 method to convert all cores in a CaliperMatrix to R matrices}
\usage{
\method{as.matrix}{CaliperMatrix}(x, ...)
}
\arguments{
\item{x}{\code{CaliperMatrix}}

\item{...}{other arguments passed to \code{as.matrix()}}
}
\value{
A named list of R matrices.
}
\description{
To convert individual cores of a \code{CaliperMatrix}, see
\code{\link{as.matrix.MatrixCurrency}}.
}
