% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{as.matrix.CaliperMatrix}
\alias{as.matrix.CaliperMatrix}
\title{S3 method for bringing caliper matrices into R matrix format}
\usage{
\method{as.matrix}{CaliperMatrix}(x, ...)
}
\arguments{
\item{x}{A \code{CaliperMatrix} object}

\item{...}{Additional arguments passed to \code{as.matrix}. An extra argument
\code{core} can be used to specify which core to convert. See details}
}
\description{
Converts a Caliper matrix into R format allowing it to be worked on in the R
environment.
}
\details{
By default, the first core of the Caliper matrix is converted into an R matrix.
To select a different one, use the \code{core} argument like so:
\code{as.matrix(x, core = "second core")}.
}
