% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{CaliperClass}
\alias{CaliperClass}
\title{R class representing GISDK objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CaliperClass
}
\description{
This class simplifies working with GISDK objects over COM from R.
}
\details{

}
\section{Overview}{


When created, an object of this class contains a COM pointer to a sister
object in Caliper software. In addition to this reference, contained in the
\code{ref} field, there are several other fields and methods to facilitate
object manipulation. Finally, the \code{$} operator has been overloaded for
this class to make code easier to write.
}

\section{Object fields}{


A \code{CaliperClass} object has the following fields/attributes of interest:
\describe{
  \item{g_class_name}{
    The class name of the underlying GISDK object (e.g. "NLM.Model").
  }
  \item{ref}{
    This is a COM pointer and represents the object created in a Caliper
    program.
  }
  \item{info}{
    This is a simple R list that can be used to see the fields and methods
    of the underlying GISDK object.
  }
}
}

\section{Object methods}{


A \code{CaliperClass} object has several methods. While they are listed
below for completeness, you likely won't need to use the methods directly
(see examples).

\describe{
  \item{apply_gisdk_method}{
    Dispatches one of the GISDK object's methods
  }
  \item{get_gisdk_attribute(attribute)}{
    Gets the value of a named attribute of the GISDK object
  }
  \item{set_gisdk_attribute(attribute)}{
    Sets the value of a named attribute of the GISDK object
  }
}
}

\examples{
\dontrun{
# object creation (easier to use caliper::CreateObject())
obj <- CaliperClass$new("NLM.Model")

# get info about the GISDK object fields and methods
obj$info

# set an attribute/field label of the GISDK object
# (Does not modify the R object in any way)
obj$Label <- "a logit model"

# run one of the GISDK object's methods: Clear()
# (Does not modify the R object in any way)
obj$Clear()

# object methods can be chained
obj$
 Clear()$
 Read()
}
}
\keyword{datasets}
