% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{CaliperMatrix}
\alias{CaliperMatrix}
\title{CaliperMatrix class}
\description{
CaliperMatrix class

CaliperMatrix class
}
\details{
This class makes it easier to interact with Caliper matrices in R.

See \code{vignette("caliper-matrices")} for more details on interacting with
matrices. This includes info on S3 methods for bringing them into R formats
like \code{data.frames} and \code{as.matrix}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{handle}}{This is a COM pointer and represents the object in Caliper
software.}

\item{\code{cores}}{This is a named list of cores. Each element contains a
\code{\link{MatrixCurrency-class}}.}

\item{\code{indices}}{This lists the row and column indices available for the
matrix.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{row_index}}{This can be used to retrieve or set the current row index.}

\item{\code{column_index}}{This can be used to retrieve or set the current column index.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CaliperMatrix$new()}}
\item \href{#method-create_index_info}{\code{CaliperMatrix$create_index_info()}}
\item \href{#method-create_core_list}{\code{CaliperMatrix$create_core_list()}}
\item \href{#method-update_gisdk_data}{\code{CaliperMatrix$update_gisdk_data()}}
\item \href{#method-CreateCore}{\code{CaliperMatrix$CreateCore()}}
\item \href{#method-CreateIndex}{\code{CaliperMatrix$CreateIndex()}}
\item \href{#method-clone}{\code{CaliperMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CaliperMatrix} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$new(matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{Either the file path to the matrix (.mtx) or a COM pointer
to the matrix handle.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{CaliperMatrix} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_index_info"></a>}}
\if{latex}{\out{\hypertarget{method-create_index_info}{}}}
\subsection{Method \code{create_index_info()}}{
Sets the indices field of the \code{CaliperMatrix}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$create_index_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_core_list"></a>}}
\if{latex}{\out{\hypertarget{method-create_core_list}{}}}
\subsection{Method \code{create_core_list()}}{
Sets the \code{cores} field of the \code{CaliperMatrix}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$create_core_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_gisdk_data"></a>}}
\if{latex}{\out{\hypertarget{method-update_gisdk_data}{}}}
\subsection{Method \code{update_gisdk_data()}}{
Sends R data back to the matrix in Caliper software.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$update_gisdk_data(core_name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{core_name}}{\code{string} The name of the core to update.}

\item{\code{data}}{\code{matrix} of data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCore"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCore}{}}}
\subsection{Method \code{CreateCore()}}{
Create a new core in the matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$CreateCore(core_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{core_name}}{\code{string} The name of the core to create.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateIndex"></a>}}
\if{latex}{\out{\hypertarget{method-CreateIndex}{}}}
\subsection{Method \code{CreateIndex()}}{
Create a new index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$CreateIndex(
  index_name,
  current_ids,
  new_ids = NULL,
  index_type = c("both", "row", "column")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_name}}{\code{string} The name of the new index.}

\item{\code{current_ids}}{\code{vector} of the row/column IDs to include in new
index.}

\item{\code{new_ids}}{\code{vector} Use this to assign new IDs to the rows/columns
of this index if desired.}

\item{\code{index_type}}{\code{string} Whether the index can be applied to rows,
column, or (the default) both.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
