% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{CaliperMatrix}
\alias{CaliperMatrix}
\title{CaliperMatrix class}
\description{
This class makes it easier to interact with Caliper matrices in R.
}
\section{Object fields}{


A \code{CaliperMatrix} object has the following fields/attributes of interest:
\describe{
  \item{handle}{
    This is a COM pointer and represents the object in Caliper Software.
  }
  \item{cores}{
    This is a named list of cores. Each element contains a COM pointer to
    a matrix currency in Caliper software.
  }
  \item{indices}{
    This lists the row and column indices available for the matrix.
  }
  \item{row_index}{
    This can be used to retrieve or set the current row index.
  }
  \item{column_index}{
    This can be used to retrieve or set the current column index.
  }
}

See the vignette "Using caliper" for more details on interacting with
matrices. This includes info on S3 methods for bringing them into R formats
like \code{data.frames} and \code{matrices}.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CaliperMatrix$new()}}
\item \href{#method-create_matrix_cores}{\code{CaliperMatrix$create_matrix_cores()}}
\item \href{#method-clone}{\code{CaliperMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$new(matrix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_matrix_cores"></a>}}
\subsection{Method \code{create_matrix_cores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$create_matrix_cores()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaliperMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
