% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\docType{package}
\name{caliperR}
\alias{caliperR}
\title{caliperR: Communicate with Caliper software from R}
\description{
The caliperR package provides functions to make it easier to communicate
with Caliper software over COM.
}
