% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{$.GisdkObject}
\alias{$.GisdkObject}
\title{S3 generic for calling GISDK object methods}
\usage{
\method{$}{GisdkObject}(x, name)
}
\arguments{
\item{x}{A \code{GisdkObject}}

\item{name}{the method to dispatch}
}
\description{
This function allows you to use \code{obj$method_name(...)} to dispatch
arbitrary method names on GISDK objects. See \code{\link{GisdkObject}}.
}
