% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{connect}
\alias{connect}
\title{Create a connection to a Caliper software product}
\usage{
connect(software = NULL, silent = FALSE)
}
\arguments{
\item{software}{One of either "TransCAD", "TransModeler", or "Maptitude". If
left \code{NULL}, the function will search (in that order) and create the first
connection it can.}

\item{silent}{\code{boolean} Whether to display a connected message.}
}
\value{
Nothing. Sets the COM object to a global environment variable
  (\code{CALIPER_DK})
}
\description{
Will create a connection over COM to one of Caliper's installed software
products. (The software must be installed with a valid license.)
}
