% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{create_matrix}
\alias{create_matrix}
\title{Creates an S3 object of class \code{CaliperMatrix}}
\usage{
create_matrix(m)
}
\arguments{
\item{m}{Either a file name to read (.mtx) or a COM pointer to an open matrix
in Caliper software.}
}
\description{
A \code{CaliperMatrix} is different from an R matrix. This object represents
all the cores of a Caliper matrix.
}
\keyword{internal}
