% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{create_view}
\alias{create_view}
\title{Creates a new Caliper view with data from a data.frame;}
\usage{
create_view(df)
}
\arguments{
\item{df}{\code{data.frame} The data to send to a new view}
}
\description{
The data view is a "MEM" (or memory table), and is not associated with
a file.
}
\keyword{internal}
