% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliperR.R
\name{process_gisdk_args}
\alias{process_gisdk_args}
\title{Convert R arguments into GISDK flavors}
\usage{
process_gisdk_args(...)
}
\arguments{
\item{arg_list}{\code{list} of args that are converted.}
}
\description{
It calls \code{\link{convert_to_named_array}} and
\code{\link{convert_nulls_and_slashes}} as appropriate on each argument passed.
}
\keyword{internal}
