% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{read_bin_without_com}
\alias{read_bin_without_com}
\title{Read the bin matrix into a data table without a COM connection}
\usage{
read_bin_without_com(binFilename, dcbKey, nBytesPerRow)
}
\arguments{
\item{binFilename}{\code{string} Path of bin file to read.}

\item{dcbKey}{\code{data.frame} The FFB table structure df. Created by
read_bin.}

\item{nBytesPerRow}{\code{numeric} The number of bytes per row.}
}
\description{
Helper function to \code{\link{read_bin}}. If Caliper software is not
installed, this function reads and translates the bytes directly.
}
\keyword{internal}
