% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{summary.CaliperMatrix}
\alias{summary.CaliperMatrix}
\title{S3 method for summarizing a \code{CaliperMatrix}}
\usage{
\method{summary}{CaliperMatrix}(object, ...)
}
\arguments{
\item{object}{A \code{CaliperMatrix} object.}

\item{...}{Additional arguments passed to the \code{summary} generic
function.}
}
\value{
A \code{data.frame} of summary statistics. These summary statistics
  are always for the full Caliper matrix and do not respect the current
  index.

A \code{data.frame} in with a row for each core.
}
\description{
Statistics return include things like sum, mean, min, etc.
}
