% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{summary.CaliperMatrix}
\alias{summary.CaliperMatrix}
\title{S3 method for summarizing a \code{CaliperMatrix}}
\usage{
\method{summary}{CaliperMatrix}(x, ...)
}
\arguments{
\item{x}{\code{CaliperMatrix}}

\item{...}{Additional arguments (not used)}
}
\description{
S3 method for summarizing a \code{CaliperMatrix}
}
