% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{update_view}
\alias{update_view}
\title{Updates an existing Caliper view with data from a data.frame}
\usage{
update_view(df, view_name, set_name = NULL)
}
\arguments{
\item{df}{\code{data.frame} The data to update the view with.}

\item{view_name}{\code{string} The view to update.}

\item{set_name}{\code{string} Optional selection set name. If provided,
only the rows within the selection set of \code{view_name} will be updated.}
}
\value{
\code{string} The view name.
}
\description{
Updates an existing Caliper view with data from a data.frame
}
\keyword{internal}
