% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin.R
\name{write_bin_using_com}
\alias{write_bin_using_com}
\title{Uses Caliper software over COM to write a bin file.}
\usage{
write_bin_using_com(binData, binFilename)
}
\arguments{
\item{binData}{\code{data.frame} The data table to write}

\item{binFilename}{\code{string} The file in which to write}
}
\description{
When available, this approach is faster than writing line by line from R.
Helper function for \code{\link{write_bin}}.
}
\keyword{internal}
